fix: bug with NewGroupModal, bigLetters in folderName, architecture mistakes, newGroupModal/button/icon ; change: newGroupModal/validations
