+++ package.json
@@ -11,6 +11,7 @@
     "dependencies": {
         "@emotion/react": "^11.11.1",
         "@emotion/server": "^11.11.0",
+        "@skbkontur/icons": "^1.7.1",
         "@skbkontur/react-icons": "^5.2.9",
         "@skbkontur/react-ui": "^4.17.3",
         "@types/node": "20.5.1",
