+++ src/app/(routing)/layout.tsx
@@ -4,7 +4,9 @@ export default function Layout({ children }: { children: React.ReactNode }) {
     return (
         <div style={{ display: 'flex', height: '100vh' }}>
             <Header />
-            <main style={{ flexGrow: '1', overflowY: 'auto' }}>{children}</main>
+            <main style={{ flexGrow: '1', overflowY: 'auto', paddingBottom: '2rem' }}>
+                {children}
+            </main>
         </div>
     );
 }
