--- src/app/(routing)/section/create/Create.tsx
@@ -1,7 +1,0 @@
-export default function Create({ params }: { params: { id: number } }) {
-    return (
-        <>
-            <newGroupModal />
-        </>
-    );
-}
