+++ src/app/(routing)/section/create/page.tsx
@@ -1,3 +1,9 @@
-export default function Page() {
-    return <>Section/Create</>;
+import { NewGroupModal } from '@/widgets/newGroupModal';
+
+export default function Create({ params }: { params: { id: number } }) {
+    return (
+        <>
+            <NewGroupModal />
+        </>
+    );
 }
