--- src/entities/Divider/Divider.style.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledDivider = styled.div`
-    width: 100%;
-    height: 1px;
-    background-color: #e6e6e6;
-`;
