--- src/entities/Divider/Divider.tsx
@@ -1,8 +1,0 @@
-import React, { FC } from 'react'
-import { StyledDivider } from './Divider.style'
-
-export const Divider: FC = () => {
-  return (
-    <StyledDivider/>
-  )
-}
