+++ src/entities/Divider/index.tsx
@@ -1,1 +1,1 @@
-export { Divider } from './Divider'
+export { Divider } from './ui/ui';
