+++ src/entities/Divider/ui/ui.module.scss
@@ -1,0 +1,5 @@
+.divider {
+    width: 100%;
+    height: 1px;
+    background-color: #e6e6e6;
+}
