+++ src/entities/Divider/ui/ui.tsx
@@ -1,0 +1,9 @@
+import styles from './ui.module.scss';
+
+export const Divider = () => {
+    return (
+        <>
+            <span className={styles.divider}></span>
+        </>
+    );
+};
