--- src/entities/Text/Text.constants.ts
@@ -1,21 +1,0 @@
-export const H1_TEXT = 'h1';
-export const H2_TEXT = 'h2';
-export const H3_TEXT = 'h3';
-export const H4_TEXT = 'h4';
-export const H5_TEXT = 'h5';
-export const H6_TEXT = 'h6';
-export const P_TEXT = 'p';
-export const SPAN_TEXT = 'span';
-
-
-export const TEXT_TYPES = [
-    H1_TEXT,
-    H2_TEXT,
-    H3_TEXT,
-    H4_TEXT,
-    H5_TEXT,
-    H6_TEXT,
-    P_TEXT,
-    SPAN_TEXT
-] as const;
-
