+++ src/entities/Text/index.ts
@@ -1,1 +1,1 @@
-export { Text } from './Text'
+export { Text } from './ui/ui';
