+++ src/widgets/NewGroupModal/ui/ui.module.scss
@@ -1,4 +1,3 @@
 .modalWrap {
     width: 518px;
-    // padding: 24px 32px;
 }
