+++ src/widgets/sectionView/ui/ui.tsx
@@ -12,10 +12,14 @@ import { GroupCard } from '@/features/groupCard';
 import { GroupItems, RatingItems } from '../data';
 import { GroupCreateCard } from '@/features/createCard';
 import { SectionRating } from '@/features/sectionRating';
+import { NewGroupModal } from '@/widgets/newGroupModal';
+import { useState } from 'react';
 export const SectionView = ({ id }: { id: number }) => {
     const router = useRouter();
+    const [modalOpen, setModalOpen] = useState<boolean>(false);
     return (
         <>
+            {modalOpen && <NewGroupModal setModalOpen={() => setModalOpen(false)} />}
             <Layout>
                 <section className={styles.header}>
                     <Title title="Вернуться на предыдущую страницу" onClick={() => router.back()}>
@@ -49,7 +53,10 @@ export const SectionView = ({ id }: { id: number }) => {
                         {GroupItems.map((item) => (
                             <GroupCard key={item.id} item={item} />
                         ))}
-                        <GroupCreateCard title="Добавить группу" />
+                        <GroupCreateCard
+                            onClick={() => setModalOpen(true)}
+                            title="Добавить группу"
+                        />
                     </div>
                 </section>
                 <section className={styles.groups}>
